requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
prettytable==2.2.1
matplotlib==3.4.3
Pillow==8.3.2

